part 2 of making shared object out of dssim
I split up the main function so that the core of it could be accessible. Still
returning retval from calculations, but moving results into an int array to be
printed later rather than printing on each iteration of the loop.

Also made choice to not print any results if encountering a bad image. Can play
around with that, though (should retval still be nonzero? etc.)